% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_contingency_table.R
\name{calculate_contingency_table}
\alias{calculate_contingency_table}
\title{Calculate contingency table.}
\usage{
calculate_contingency_table(medications, covariates)
}
\arguments{
\item{medications}{A tibble}

\item{covariates}{A tibble}
}
\value{
A tibble \code{cont_table_output}
}
\description{
Calculate contingency table.
}
