% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_regression_table.R
\name{create_regression_table}
\alias{create_regression_table}
\title{Create regression table.}
\usage{
create_regression_table(drug_allergies, covariates)
}
\arguments{
\item{covariates}{A tibble}

\item{drug_allergie}{A tibble}
}
\value{
A tibble \code{regression_table}
}
\description{
Create regression table.
}
