% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiwas_regression.R
\name{ddiwas_regression}
\alias{ddiwas_regression}
\title{Logistic regression function.}
\usage{
ddiwas_regression(regression_table, rare_events = FALSE)
}
\arguments{
\item{regression_table}{A tibble}

\item{rare_events}{Binary value. Defaults to FALSE. Change to TRUE if you want to test drugs with rare ADR events}
}
\value{
A tibble \code{results}
}
\description{
Logistic regression function.
}
