% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_drug_allergies.R
\name{map_drug_allergies}
\alias{map_drug_allergies}
\title{Map drug allergies to RxCUI ingredients.}
\usage{
map_drug_allergies(ehr_drug_allergies)
}
\arguments{
\item{ehr_drug_allergies}{A tibble dataframe}
}
\value{
A tibble dataframe \code{drug_allergies}
}
\description{
Map drug allergies to RxCUI ingredients.
}
