% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_covariates.R
\name{process_covariates}
\alias{process_covariates}
\title{Process covariates.}
\usage{
process_covariates(covariates, df_drug_exposures)
}
\arguments{
\item{covariates}{A tibble dataframe}

\item{df_drug_exposures}{A tibble dataframe}
}
\value{
A tibble dataframe \code{covariates}
}
\description{
Process covariates.
}
