% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_drug_allergies.R
\name{require_drug_allergies}
\alias{require_drug_allergies}
\title{Removes drug ingredient allergies documented prior to observation period and requires all patients to have >=1 drug allergy list entry during observation period.}
\usage{
require_drug_allergies(ehr_drug_allergies, covariates)
}
\arguments{
\item{ehr_drug_allergies}{A tibble}

\item{covariates}{A tibble}
}
\value{
A tibble \code{covariates}
}
\description{
Removes drug ingredient allergies documented prior to observation period and requires all patients to have >=1 drug allergy list entry during observation period.
}
